clients/habibi/requirements.txt
crewai>=0.28.0
python-dotenv>=1.0.0
openai>=1.0.0


